Force 2fa for admin users
The Jetpack filter to force 2fa is applied before the local user object has been set, so we need to mimic Jetpack's process of obtaining the local user from the wpcom_id or wpcom email address.

Adds a function the hook just prior to the 2fa check and duplicates an existing JP function pending upstream patch to make static, not yet submitted/discused with JP devs.