Load Jetpack Force 2FA later
On VIP Go, we enabled this plugin on `setup_theme` because there may be some theme settings that factor into whether the plugin is enabled or not. We therefore can't force 2fa before after_setup_theme. This should be fine in all cases because the theme is set up before logins happen.